Move Footer to /complex
