+++ packages/york-web/src/components/complex/index.js
@@ -1,4 +1,5 @@
 export { default as CallToAction } from './CallToAction'
 export { default as ErrorPage } from './ErrorPage'
+export { default as Footer } from './Footer'
 export { default as Header } from './Header'
 export { default as PromoCard } from './PromoCard'
