--- packages/york-web/src/components/simple/Footer/README.md
@@ -1,22 +1,0 @@
-```js
-;<div style={{ overflow: 'auto'}}>
-  <Footer
-    legalInfo="© 2019, Qlean"
-    email="feedback@qlean.ru"
-    userAgreementLink=""
-    phones={['88005008264', '84953747907']}
-    mobileStoreLinks={{
-      googlePlay:
-        'https://apps.apple.com/ru/app/qlean-%D1%83%D0%B1%D0%BE%D1%80%D0%BA%D0%B0-%D0%BA%D0%B2%D0%B0%D1%80%D1%82%D0%B8%D1%80/id1011771034',
-      appStore:
-        'https://play.google.com/store/apps/details?id=com.qlean.qlean&hl=ru',
-    }}
-    socialNetworkLinks={{
-      vk: '',
-      facebook: '',
-      telegram: '',
-      viber: '',
-    }}
-  />
-</div>
-```
