+++ packages/york-web/src/components/simple/index.js
@@ -1,5 +1,6 @@
+export { default as Accordion } from './Accordion'
 export { default as Button } from './Button'
-export { default as Footer } from './Footer'
 export { default as GridColumn } from './GridColumn'
 export { default as GridContainer } from './GridContainer'
+export { default as SocialButton } from './SocialButton'
 export { default as Tooltip } from './Tooltip'
